refactor [James]: refactor the mail template so that it will be reusable
