+++ server-side/controllers/index.controller.js
@@ -4,6 +4,7 @@ const bcrypt = require("bcryptjs");
 const Student = require("../models/StudentModel");
 const Admin = require("../models/AdminModel");
 const Log = require("../models/LogModel");
+const { forgotPasswordMail } = require("../mail_template/mail.template");
 require("dotenv").config();
 const token_key = process.env.JWT_SECRET;
 
@@ -160,19 +161,19 @@ const forgotPasswordController = async (req, res) => {
       user = getUser;
     } else if (!getUser) {
       console.error(`User with email ${req.body.email} not found`);
-      return res.status(404).send({
+      return res.status(404).json({
         message: `The id number you entered is found but appears to be the email is incorrect.`,
       });
     } else if (!userAdmin) {
       console.error(
         `The id number you entered is found but appears to be the email is incorrect.`
       );
-      return res.status(404).send({
+      return res.status(404).json({
         message: `The id number you entered is found but appears to be the email is incorrect.`,
       });
     } else {
       console.error(`User with email ${req.body.email} not found`);
-      return res.status(404).send({
+      return res.status(404).json({
         message: `The id number you entered is found but appears to be the email is incorrect.`,
       });
     }
@@ -180,56 +181,7 @@ const forgotPasswordController = async (req, res) => {
     const token = jwt.sign({ userId: user._id }, process.env.JWT_SECRET, {
       expiresIn: "10m",
     });
-
-    const transporter = nodemailer.createTransport({
-      service: "gmail",
-      auth: {
-        user: process.env.EMAIL,
-        pass: process.env.PASSWORD_APP_EMAIL,
-      },
-    });
-
-    const mailOptions = {
-      from: process.env.EMAIL,
-      to: req.body.email,
-      subject: "Reset Your Password",
-      html: `
-        <div style="font-family: Arial, sans-serif; max-width: 600px; margin: 0 auto; padding: 20px; border: 1px solid #ddd; border-radius: 10px;">
-          <h1 style="color: #333; text-align: center; margin-bottom: 30px;">PSITS - Reset Your Password</h1>
-          <p style="color: #555; font-size: 16px;">Hello,</p>
-          <p style="color: #555; font-size: 16px; margin-bottom: 20px">
-            We received a request to reset your password. Click the button below to reset it:
-          </p>
-          <div style="text-align: center; margin: 40px 0;">
-            <a
-              href="${url}${token}" 
-              style="display: inline-block; padding: 20px 25px; color: #fff; background-color: #007bff; text-decoration: none; border-radius: 5px; font-size: 24px;">
-              Reset Password
-            </a>
-          </div>
-          <p style="color: #555; font-size: 16px;">Or you can copy and paste this link into your browser:</p>
-          <p style="word-break: break-all;">
-            <a href="${url}${token}" style="color: #007bff;">
-             ${url}${token}
-            </a>
-          </p>
-          <p style="color: #999; font-size: 14px;">
-            This link will expire in 10 minutes. If you didn’t request a password reset, you can safely ignore this email.
-          </p>
-          <p style="color: #555; font-size: 16px;">Thank you,</p>
-          <p style="color: #555; font-size: 16px;">The Support Team</p>
-        </div>
-      `,
-    };
-
-    transporter.sendMail(mailOptions, (err, info) => {
-      if (err) {
-        console.error("Error sending email:", err.message);
-        return res.status(500).send({ message: err.message });
-      }
-      //console.log("Email sent from forgot password:", info.response);
-      res.status(200).send({ message: "Email sent" });
-    });
+    await forgotPasswordMail(req.body.email, url, token);
   } catch (err) {
     console.error("Server error during forgot password process:", err.message);
     res.status(500).send({ message: err.message });
